\name{pima.indians.diabetes}
\alias{pima.indians.diabetes}
\title{Pima Indians Diabetes Database}
\usage{data(pima.indians.diabetes)}
\keyword{datasets}
\format{
    A data frame with 768 observations on 9 variables:
    \tabular{rl}{
   1 \tab Number of times pregnant\cr
   2 \tab Plasma glucose concentration (glucose tolerance test)\cr
   3 \tab Diastolic blood pressure (mm Hg)\cr
   4 \tab Triceps skin fold thickness (mm)\cr
   5 \tab 2-Hour serum insulin (mu U/ml)\cr
   6 \tab Body mass index (weight in kg/(height in m)^2)\cr
   7 \tab Diabetes pedigree function\cr
   8 \tab Age (years)\cr
   9 \tab Class variable (test for diabetes)\cr
  }
}
\source{
    \itemize{
	\item Original owners: National Institute of Diabetes and Digestive and
        Kidney Diseases
	\item Donor of database: Vincent Sigillito
	(vgs@aplcen.apl.jhu.edu)
    }
    
    These data have been taken from the UCI Repository Of Machine Learning
    Databases at
    \itemize{
    	\item ftp.ics.uci.edu://pub/machine-learning-databases
    	\item http://www.ics.uci.edu/mlearn/MLRepository.html
    }
    and were converted to R format by Friedrich.Leisch@ci.tuwien.ac.at.
}
    
