\name{glass}
\alias{glass}
\title{Glass Identification Database}
\usage{data(glass)}
\keyword{datasets}
\format{
    A data frame with 214 observations on 11 variables:
    \tabular{cll}{
 [,1] \tab Number \tab ID number\cr
 [,2] \tab RI \tab refractive index\cr
 [,3] \tab Na \tab Sodium\cr
 [,4] \tab Mg \tab Magnesium\cr
 [,5] \tab Al \tab Aluminum\cr
 [,6] \tab Si \tab Silicon\cr
 [,7] \tab K  \tab Potassium\cr
 [,8] \tab Ca \tab Calcium\cr
 [,9] \tab Ba \tab Barium\cr
[,10] \tab Fe \tab Iron \cr
[,11] \tab Type \tab Type of glass (class attribute) \cr 
}
}   
\source{
    \itemize{
       	\item Creator: B. German, Central Research Establishment, Home
	Office Forensic Science Service, Aldermaston, Reading, Berkshire
	RG7 4PN 
   	\item Donor: Vina Spiehler, Ph.D., DABFT, Diagnostic Products
	Corporation
    }
    
    These data have been taken from the UCI Repository Of Machine Learning
    Databases at
    \itemize{
    	\item ftp.ics.uci.edu://pub/machine-learning-databases
    	\item http://www.ics.uci.edu/mlearn/MLRepository.html
    }
    and were converted to R format by Friedrich.Leisch@ci.tuwien.ac.at.
}
